requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
lxml==4.4.2
pkg-resources==0.0.0
requests==2.22.0
smmap2==2.0.5
soupsieve==1.9.5
urllib3==1.25.7
